navigation view u fragmentu prikazuje ok brojeve pitanja

+++ app/src/main/java/ba/etf/rma21/projekat/data/models/PitanjeKviz.kt
@@ -4,6 +4,6 @@ data class PitanjeKviz(
         val naziv: String,
         val kviz: String,
         val predmetNaziv: String,
-        val bodovi: Double
+        val bodovi: Float
 ) {
 }
